Refactor node helpers into nodes module
